+++ include/positionedNodeSet.h
@@ -43,7 +43,7 @@ public:
     /// Return a string representation of the NodeSet
     std::string toString();
     /// Return the number of particles at each Node in the NodeSet
-    std::vector<unsigned int> numberOfParticles();
+    std::vector<unsigned long long> numberOfParticles();
     /**
      * Make one time step
      *
@@ -57,7 +57,7 @@ public:
 	 */
 	void reinitialize();
     ///Get positions
-    std::unordered_map<unsigned int, std::array<double, 2>> getPositions();
+    std::unordered_map<unsigned long long, std::array<double, 2>> getPositions();
     //Get PositionedNodes
     std::vector<std::shared_ptr<PositionedNode<2>>> getNodes();
     
@@ -93,7 +93,7 @@ public:
 	 * @param id The in-memory id of the Node
 	 * @return The file id of the Node
 	 */
-	int getFileId(unsigned int id) {
+	long long getFileId(unsigned long long id) {
 		return inverseIdMap[id];
 	}
 	/**
@@ -102,7 +102,7 @@ public:
 	 * @param id The file id of the Node
 	 * @return The in-memory id of the Node
 	 */
-	unsigned int getMemoryId(int id) {
+	unsigned long long getMemoryId(long long id) {
 		return idMap[id];
 	}
 	/**
@@ -112,7 +112,7 @@ public:
 	 * @param source The source Node
 	 * @return A map containing the shortest path from source to each node (in-memory node id is the key)
 	 */
-	std::pair<std::unordered_map<unsigned int, double>, std::unordered_map<unsigned int, int>> shortestPath(unsigned int source);
+	std::pair<std::unordered_map<unsigned long long, double>, std::unordered_map<unsigned long long, long long>> shortestPath(unsigned long long source);
 	/**
 	 * Write this set in METIS format.
 	 *
@@ -135,14 +135,14 @@ public:
 	 * @param filename The file to write to
 	 * @param numPart The number of partitions used by METIS
 	 */
-   void readMETIS(std::istream& input, std::ostream& output, unsigned int numPart);
+   void readMETIS(std::istream& input, std::ostream& output, unsigned long long numPart);
 	/**
 	 * Reorder nodes and write to a file using a METIS partition.
 	 *
 	 * @param filename The file to write to
 	 * @param numPart The number of partitions used by METIS
 	 */
-   void readMETIS(const std::string& input, const std::string& output, unsigned int numPart) {
+   void readMETIS(const std::string& input, const std::string& output, unsigned long long numPart) {
         std::ifstream in(input);
         std::ofstream out(output);
         readMETIS(in, out, numPart);
@@ -191,9 +191,9 @@ private:
     ///A vector containing all Algorithms in the set
     std::vector<std::shared_ptr<StorableAlgorithm>> algorithms;
     ///Map from the ids in the input stream to the actual ids of the Nodes
-    std::unordered_map<int, unsigned int> idMap;
+    std::unordered_map<long long, unsigned long long> idMap;
     ///Map from the actual ids of the Nodes to the ids in the input stream 
-    std::unordered_map<unsigned int, int> inverseIdMap;
+    std::unordered_map<unsigned long long, long long> inverseIdMap;
     ///Helper function to initialize algorithms
     ///@param create The function with which to create the Algorithm for each Node
     ///@param element The Element that the Algorithm should have
